Move Game code in a separate file
